mgr: store names of modules that register RADOS clients in the MgrMap
The MgrMap stores a list of RADOS clients addresses registered by the
mgr modules. During failover of ceph-mgr, the list is used to blocklist
clients belonging to the failed ceph-mgr.

In the MgrMap, store the names of the mgr modules that registered the
RADOS clients. This allows one to easily identify the mgr module that
registered a particular RADOS client by just dumping the MgrMap
(`ceph mgr dump`).

Following is the MgrMap output with a module's client name displayed
along with its client addrvec,
$ ceph mgr dump | jq '.always_on_modules.active_clients[0]'
{
  "name": "devicehealth",
  "addrvec": [
    {
      "type": "v2",
      "addr": "10.0.0.148:0",
      "nonce": 612376578
    }
  ]
}

Signed-off-by: Ramana Raja <rraja@redhat.com>