qa: Add tests to validate synced images on rbd-mirror
Introduce integration tests to validate the checksums of images synced
using snapshot based rbd mirroring.

Run workload on an image using a krbd or nbd client. Take mirror
snapshots of the image under workload. Unmap and demote the primary
image. Calculate the MD5 checksum of the image's most recent mirror
snapshot. Wait for the syncing of the mirror snapshot to complete.
Promote the non-primary image in the other cluster. Calculate the
MD5 checksum of the newly promoted image. Check whether the hashes of
the demoted image's and promoted image's latest mirror snapshots are
the same.

The above test is run as part of two different workunits:
 - a workunit that validates the syncing of multiple images with
   workloads running on them
 - another workunit that validates the syncing of a single image with
   workload running on it and the primary image alternates between
   the clusters.

Fixes: https://tracker.ceph.com/issues/61617
Signed-off-by: Christopher Hoffman <choffman@redhat.com>
Co-authored-by: Ramana Raja <rraja@redhat.com>